+++ OpenGL/OpenGL/src/entity/Entity.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "Hitbox.hpp"
+#include "Body.hpp"
+#include "../audio/SoundEffect.hpp"
+
+class Entity
+{
+public:
+	Entity();
+	virtual ~Entity();
+	virtual void move();
+
+private:
+
+};
